Submitting a story with “Negative votes to remove submission” set to a number other than zero.
If we set the value of “Negative votes to remove submission” in Admin Panel -> Settings -> Voting to any number other than zero, all submitted stories after this change will be discarded. The code is inaccurately processing the “buries_to_spam” variable which reflects the value set above.

To demonstrate how it is executed, I will show the trace of the PHP statements being called and executed:

In \submit.php, inside the Function do_submit1, a call to store in the links table what was entered in this step before moving to the next step do_submit2. We find on LINE 246 a call to $linkres->store();

In \libs\link.php, within the function store(), a call $this->store_basic(); is made on LINE 141

Now in Function store_basic(), we find the following statement of LINE 208. (Comments in red are mine to explain the process):

See the trace :
http://mymonalisasmile.com/upload/submitting-a-story-with-Negative-votes-to-remove-submission-other-than-zero.html