Second Step
In Admin Panel -> Settings -> Voting -> Negative votes to remove submission is set to 0 by default. However, in the current code, this defined variable that is assigned to buries_to_spam does not achieve its goal to set the story status to 'discard' when it reaches the number we set in here. As a matter of fact, even with the default value of 0, the story will be sent to discard. The fact is that the function evaluate_formulas () in \libs\link.php is what decides the status of the story when the formula set in the table_formulas is evaluated.

See the complete trace:

http://mymonalisasmile.com/upload/Negative-votes-to-remove-submission.html